Expanded matrix row doesn't support drag and drop #10597
